app/services/db.py
import os
from pymongo import MongoClient
from dotenv import load_dotenv

load_dotenv()

MONGO_URI = os.getenv("MONGO_URI")

# Connexion MongoDB
client = MongoClient(MONGO_URI)
db = client.get_database()

# Collections
patients_col = db["patients"]
analyses_col = db["analyses"]
